version 2.0.4.1 fixed server not starting fixed unlimited speed when floating on water. Nerfed DragonUpgrade. Added config threshold for better controller support.
