+++ src/main/resources/META-INF/mods.toml
@@ -14,7 +14,7 @@ loaderVersion="[32,)" #mandatory This is typically bumped every Minecraft versio
 # The modid of the mod
 modId="simpleplanes" #mandatory
 # The version number of the mod - there's a few well known ${} variables useable here or just hardcode it
-version="2.0.4.0" #mandatory
+version="2.0.4.1" #mandatory
  # A display name for the mod
 displayName="Simple Planes" #mandatory
 # A URL to query for updates for this mod. See the JSON update specification <here>
